--- a/activesupport/lib/active_support/dependencies/autoload.rb
+++ b/activesupport/lib/active_support/dependencies/autoload.rb
@@ -79,4 +79,4 @@ module ActiveSupport
       @_inflector_loaded = true
     end
   end
-end 
+end