### Motivation / Background
While benchmarking and profiling the rails library i noticed there was around a 79.83 ms bottleneck that caused by this single line in the original autoload file:
```
require "active_support/inflector/methods"
```

We could implement a lazy loading mechanism, and loading when it is required.

### Detail
Defer heavy inflector loading until needed, addresses bottleneck in ActiveSupport initialization profiling.

### Additional information

Reasoning behind(profiling steps):
- Initial profiling identified `gem loading` as a potential candidate for an optimization
- Component profiling found `ActiveSupport` was taking 146.24 ms
- Individual `require` profiling discovered `autoload.rb` was taking 79.83 ms
- Found `require active_support/inflector/methods` was the heavy loading line

This should reduce the bottleneck significantly, `autoload` initialization from 79.83ms to 1.57ms

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [x] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.