@jonbarlo 2025-08-28
@matthewd hello ! any particular reason why you closed this one? i see no feedback just a closure, this one will speed things up i already benchmark and profile it.